pyproject.toml
[project]
name = "ai-blog-writer"
version = "0.1.0"
description = "AI Blog Writer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-api-python-client>=2.169.0",
    "google-auth>=2.39.0",
    "google-auth-oauthlib>=1.2.2",
    "google-search-results>=2.4.2",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-google-genai>=2.1.4",
    "loguru>=0.7.3",
    "markdown>=3.8",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "serpapi>=0.1.5",
    "uuid>=1.30",
    "validators>=0.35.0",
]
